Fix issues reviewed
Fix issues reviewed
