+++ helper/helper.go
@@ -1,0 +1,12 @@
+package helper
+
+import (
+	"crypto/md5"
+	"encoding/hex"
+)
+
+func GetMD5Hash(text string) string {
+	hasher := md5.New()
+	hasher.Write([]byte(text))
+	return hex.EncodeToString(hasher.Sum(nil))
+}
