+++ models/user.go
@@ -15,16 +15,3 @@ type User struct {
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"update_at"`
 }
-type TransformUser struct {
-	ID          int    `json:"id"`
-	Username    string `json:"username"`
-	Email       string `json:"email"`
-	Type        int    `json:"type"`
-	Gender      string `json:"gender"`
-	BirthDay    string `json:"birthday"`
-	PhoneNumber int    `json:"phone_number"`
-	Status      int    `json:"status"`
-}
-type ListUser struct {
-	Members []TransformUser
-}
